Simplified the upgrade process by leveraging automated draining of nodes.
